rename interface file
