+++ WebApp/webapp/src/components/MatchupSlider/index.tsx
@@ -2,7 +2,7 @@ import React, { ChangeEvent, useState } from 'react';
 import styles from './index.module.css';
 import { Grid, Slider, ThemeProvider, createMuiTheme, createTheme, makeStyles } from '@mui/material';
 
-import { MatchupData } from '../../interfaces/data';
+import { MatchupData } from '../../interfaces/MatchupData';
 import { statsMap } from '../../constants/statDictionary';
 import { multby100 } from '../../util/math';
